requirements.txt
fastapi
pydantic
requests
uvicorn
python-multipart
aiofiles
python-dotenv
discord.py
deta